Updating repo structure and documentation
