+++ C2/README.md
@@ -1,0 +1,3 @@
+# C2 Overview
+
+This folder contains the code for the C2s and is currently separated by the radio type.
