+++ Implant/PCB_Designs/IH-AAA-NNC-XXX/README.md
@@ -1,0 +1,12 @@
+# IH-AAA-NNC-XXX-\<Revison\>
+These files are the EASYEDA files for the electrical schematic and the PCB layout for our initial release.  
+This board uses external crystals.
+This board is not easily compatible with an enclosure.  
+USB1 is the input from the keyboard.\*  
+USB2 is the output from the implant to target.\*
+
+\*\*\***NOTE:**\*\*\*
+- The original board that was given away in Las Vegas (marked REV 2.30 on the back) had some incorrect labeling
+- USB1 and USB2 are switched. On this board USB2 handles the keyboard input and USB1 sends out the keystrokes
+- P1 and P8 on the back of the board are next to the radio connector instead of the SWD pins located right below it
+- These labeling issues will be corrected in future releases
