+++ Implant/PCB_Designs/IH_AAA_NEA_XXX/README.md
@@ -1,0 +1,6 @@
+# IH-AAA-NEA-XXX-\<Revison\>
+These files are the EASYEDA files for the electrical schematic and the PCB layout for our initial release.  
+This board does not use external crystals.
+This board is compatible with an enclosure!  The enclosure files are incldued in the Enclosure folder.  
+USB1 is the input from the keyboard.  
+USB2 is the output from the implant to target.
