+++ Implant/PCB_Designs/IH_AAA_NNA_XXX/README.md
@@ -1,0 +1,5 @@
+# IH-AAA-NNA-XXX-\<Revison\>
+These files are the EASYEDA files for the electrical schematic and the PCB layout for our initial release.  
+This board does not use external crystals and does not allow for convenient enclosure mounting.  
+USB1 is the input from the keyboard.  
+USB2 is the output from the implant to target.
