+++ Implant/Software/IH-AAA_to_C/README.md
@@ -1,0 +1,4 @@
+# IH-AAA/C Code
+This folder contains the code for all boards using two SAMD21 chips, any XBee radio, and the Micro SD card.  
+MC1 code handles the keystroke input from the keyboard.  
+MC2 code handles keystroke recording, SD card management, keystroke injection, etc.
