+++ Implant/Software/SAMD21_Firmware/README.md
@@ -1,0 +1,5 @@
+# SAMD21 Firmware
+There are currently two variations of firmware for the SAMD21 chips.
+A GitHub link for the firmware can be found [here](https://github.com/arduino/ArduinoCore-samd/tree/master/bootloaders). 
+1) samd21_sam_ba.hex - this firmware goes to assembled boards handed out in Las Vegas that have the external crystals
+2) samd21_nano33iot.hex - this firmware goes to bare boards handed out in Las Vegas, and any boards built without the external crystals
