+++ Radios/README.md
@@ -1,0 +1,3 @@
+# Radios
+We currently only have support for the Digi XBee radio.  LoRa radios are currently being tested.
+Please see the corresponding folders for information on how to set up each type of radio.
